Rename actions.tsx to context.tsx, which is rather more fitting.
